+++ src/TRegx/CleanRegex/Replace/By/ByGroupReplacePattern.php
@@ -1,44 +1,34 @@
 <?php
 namespace TRegx\CleanRegex\Replace\By;
 
-use TRegx\CleanRegex\Exception\GroupNotMatchedException;
 use TRegx\CleanRegex\Exception\MissingReplacementKeyException;
+use TRegx\CleanRegex\Replace\GroupReplace;
 
-interface ByGroupReplacePattern
+interface ByGroupReplacePattern extends GroupReplace
 {
-    public function orElseThrow(string $exceptionClassName = GroupNotMatchedException::class): string;
-
-    public function orElseIgnore(): string;
-
-    public function orElseEmpty(): string;
-
-    public function orElseWith(string $replacement): string;
-
-    public function orElseCalling(callable $replacementProducer): string;
-
     /**
      * @param string[] $occurrencesAndReplacements
-     * @return OptionalStrategySelector
+     * @return UnmatchedGroupStrategy
      * @throws \InvalidArgumentException
      * @throws MissingReplacementKeyException
      */
-    public function map(array $occurrencesAndReplacements): OptionalStrategySelector;
+    public function map(array $occurrencesAndReplacements): UnmatchedGroupStrategy;
 
     /**
      * @param string[] $occurrencesAndReplacements
-     * @return OptionalStrategySelector
+     * @return UnmatchedGroupStrategy
      * @throws \InvalidArgumentException
      */
-    public function mapIfExists(array $occurrencesAndReplacements): OptionalStrategySelector;
+    public function mapIfExists(array $occurrencesAndReplacements): UnmatchedGroupStrategy;
 
     public function callback(callable $callback): string;
 
     /**
      * @param string[] $occurrencesAndReplacements
      * @param callable $mapper
-     * @return OptionalStrategySelector
+     * @return UnmatchedGroupStrategy
      * @throws \InvalidArgumentException
      * @throws MissingReplacementKeyException
      */
-    public function mapAndCallback(array $occurrencesAndReplacements, callable $mapper): OptionalStrategySelector;
+    public function mapAndCallback(array $occurrencesAndReplacements, callable $mapper): UnmatchedGroupStrategy;
 }