+++ src/TRegx/CleanRegex/Replace/By/UnmatchedGroupStrategy.php
@@ -4,12 +4,14 @@ namespace TRegx\CleanRegex\Replace\By;
 use TRegx\CleanRegex\Exception\GroupNotMatchedException;
 use TRegx\CleanRegex\Internal\Exception\Messages\Group\ReplacementWithUnmatchedGroupMessage;
 use TRegx\CleanRegex\Replace\GroupMapper\GroupMapper;
+use TRegx\CleanRegex\Replace\GroupReplace;
 use TRegx\CleanRegex\Replace\NonReplaced\ComputedMatchStrategy;
 use TRegx\CleanRegex\Replace\NonReplaced\ConstantReturnStrategy;
+use TRegx\CleanRegex\Replace\NonReplaced\DefaultStrategy;
 use TRegx\CleanRegex\Replace\NonReplaced\MatchRs;
 use TRegx\CleanRegex\Replace\NonReplaced\ThrowStrategy;
 
-class OptionalStrategySelectorImpl implements OptionalStrategySelector
+class UnmatchedGroupStrategy implements GroupReplace
 {
     /** @var GroupFallbackReplacer */
     private $replacer;
@@ -25,22 +27,34 @@ class OptionalStrategySelectorImpl implements OptionalStrategySelector
         $this->mapper = $mapper;
     }
 
-    public function orReturn($substitute): string
+    public function orElseWith(string $replacement): string
     {
-        return $this->replace(new ConstantReturnStrategy($substitute));
+        return $this->replace(new ConstantReturnStrategy($replacement));
     }
 
-    public function orElse(callable $substituteProducer): string
+    public function orElseCalling(callable $replacementProducer): string
     {
-        return $this->replace(new ComputedMatchStrategy($substituteProducer, "orElse"));
+        return $this->replace(new ComputedMatchStrategy($replacementProducer, "orElseCalling"));
     }
 
-    public function orThrow(string $exceptionClassName = GroupNotMatchedException::class): string
+    public function orElseThrow(string $exceptionClassName = null): string
     {
-        return $this->replace(new ThrowStrategy($exceptionClassName, new ReplacementWithUnmatchedGroupMessage($this->nameOrIndex)));
+        return $this->replace(new ThrowStrategy(
+            $exceptionClassName ?? GroupNotMatchedException::class,
+            new ReplacementWithUnmatchedGroupMessage($this->nameOrIndex)));
     }
 
-    public function replace(MatchRs $substitute): string
+    public function orElseIgnore(): string
+    {
+        return $this->replace(new DefaultStrategy());
+    }
+
+    public function orElseEmpty(): string
+    {
+        return $this->replace(new ConstantReturnStrategy(''));
+    }
+
+    private function replace(MatchRs $substitute): string
     {
         return $this->replacer->replaceOrFallback($this->nameOrIndex, $this->mapper, $substitute);
     }